+++ include/Core/host.hpp
@@ -9,6 +9,8 @@
 
 namespace panic {
 
+constexpr int MAX_HOSTNAME_LEN = 256;
+
 enum HostStatus {
   OFFLINE,
   ONLINE
@@ -39,8 +41,8 @@ public:
     auto find_port(const port& port) {return ports_.find(port.get_value());}
     // TODO write iterator adapter to port map
     // Port iterators
-    auto ports_begin() const {return ports_.begin();}
-    auto ports_end() const {return ports_.end();}
+    auto ports_begin() {return ports_.begin();}
+    auto ports_end() {return ports_.end();}
 
     // Response time manipulation
     void push_back_response_time(time_t time) {responseTime_.push_back(time);}
@@ -53,8 +55,8 @@ public:
     auto& get_last_response_time() {return responseTime_.back();}
 
     // Response time iterators
-    auto response_time_begin() const {return responseTime_.begin();}
-    auto response_time_end() const {return responseTime_.end();}
+    auto response_time_begin() {return responseTime_.begin();} 
+    auto response_time_end() {return responseTime_.end();}
 
     // Getters
     std::string get_name() const {return hostName_;}
